+++ client/src/main/java/com/reine/client/render/chunk/ChunkQuad.java
@@ -2,6 +2,7 @@ package com.reine.client.render.chunk;
 
 import com.reine.util.WorldSide;
 import org.joml.Vector3f;
+import org.joml.Vector3i;
 
-public record ChunkQuad(Vector3f start, Vector3f end, WorldSide side, int blockId) {
+public record ChunkQuad(Vector3i start, Vector3i end, WorldSide side, int blockId) {
 }
